+++ packages/graphing/configure/src/defaults.js
@@ -7,8 +7,8 @@ export default {
     answers: {
       correctAnswer: {
         name: 'Correct Answer',
-        marks: []
-      }
+        marks: [],
+      },
     },
     arrows: {
       left: true,
@@ -44,7 +44,6 @@ export default {
     rationale: '',
     standardGrid: false,
     title: '',
-    titleEnabled: true,
     toolbarTools: allTools,
     coordinatesOnHover: false,
     promptEnabled: true,
@@ -343,7 +342,8 @@ export default {
     },
     instruction: {
       settings: false,
-      label: 'Graphing questions involve plotting points, lines, or other objects on a graph. To create one, first configure the grid, then select the plotting tools students will be offered, and use them to define the correct answer.',
+      label:
+        'Graphing questions involve plotting points, lines, or other objects on a graph. To create one, first configure the grid, then select the plotting tools students will be offered, and use them to define the correct answer.',
     },
     settingsPanelDisabled: false,
     spellCheck: {