+++ packages/multiple-choice/src/choice.jsx
@@ -79,22 +79,22 @@ Choice.propTypes = {
   gridColumns: PropTypes.string,
 };
 
-export default withStyles({
+export default withStyles((theme) => ({
   choice: {
-    paddingTop: '20px',
-    paddingBottom: '10px',
-    borderBottom: '1px solid #E0DEE0',
+    paddingTop: theme.spacing.unit * 2.5,
+    paddingBottom: theme.spacing.unit + 2,
+    borderBottom: `1px solid ${theme.palette.grey[300]}`,
   },
   noBorder: {
     borderBottom: 'none',
   },
   horizontalLayout: {
-    paddingRight: '20px',
+    paddingRight: theme.spacing.unit * 2.5,
     '& label': {
-      marginRight: '8px',
+      marginRight: theme.spacing.unit,
       // '& span:first-child': {
       //   paddingRight: 0
       // }
     },
   },
-})(Choice);
+}))(Choice);