+++ packages/complex-rubric/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric@1.7.5...@pie-element/complex-rubric@1.7.7) (2023-03-20)
+
+
+### Bug Fixes
+
+* version bump ([b3f6919](https://github.com/pie-framework/pie-elements/commit/b3f6919c3ad2078953ed5162327cf356090c9aec))
+
+
+
+
+
+## [1.7.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric@1.7.4...@pie-element/complex-rubric@1.7.5) (2023-03-20)
+
+
+### Bug Fixes
+
+* update pie-libs ([340825a](https://github.com/pie-framework/pie-elements/commit/340825aab8c88569378eb5af6c5b9c9b99aeb285))
+* update pie-libs ([1fe3d37](https://github.com/pie-framework/pie-elements/commit/1fe3d37d2b001f7944e4da7c895a83ccecefefcb))
+
+
+
+
+
 ## [1.7.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/complex-rubric@1.7.3...@pie-element/complex-rubric@1.7.4) (2023-02-28)