+++ packages/protractor/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/protractor@2.9.0...@pie-element/protractor@2.9.2) (2023-03-20)
+
+
+### Bug Fixes
+
+* version bump ([b3f6919](https://github.com/pie-framework/pie-elements/commit/b3f6919c3ad2078953ed5162327cf356090c9aec))
+
+
+
+
+
+# [2.9.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/protractor@2.8.1...@pie-element/protractor@2.9.0) (2023-03-20)
+
+
+### Features
+
+* added color theming support PD-310 ([64976e0](https://github.com/pie-framework/pie-elements/commit/64976e08151b346f0760aedee2cd0e4ce6ff7f8b))
+
+
+
+
+
 ## [2.8.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/protractor@2.8.0...@pie-element/protractor@2.8.1) (2023-02-28)
 
 **Note:** Version bump only for package @pie-element/protractor